Extract rate limiting out to a concern for sharing
Previously rate limiting wasn't applied to sponsor emails which isn't
a big concern since you can only have 20 sponsors but applying the
block list to the sponsors will reduce the moderation load on the team.
