+++ app/controllers/signatures_controller.rb
@@ -144,10 +144,6 @@ class SignaturesController < ApplicationController
     unless @petition.visible?
       raise ActiveRecord::RecordNotFound, "Unable to find Signature with id: #{signature_id}"
     end
-
-    if @signature.invalidated? || @signature.fraudulent?
-      raise ActiveRecord::RecordNotFound, "Unable to find Signature with id: #{signature_id}"
-    end
   end
 
   def build_signature(attributes)
