+++ app/jobs/petition_and_email_confirmation_for_sponsor_email_job.rb
@@ -1,4 +1,6 @@
 class PetitionAndEmailConfirmationForSponsorEmailJob < EmailJob
   self.mailer = SponsorMailer
   self.email = :petition_and_email_confirmation_for_sponsor
+
+  include RateLimiting
 end
