+++ app/jobs/sponsor_signed_email_on_threshold_email_job.rb
@@ -1,4 +1,10 @@
 class SponsorSignedEmailOnThresholdEmailJob < EmailJob
   self.mailer = SponsorMailer
   self.email = :sponsor_signed_email_on_threshold
+
+  def perform(signature)
+    if signature.validated?
+      super
+    end
+  end
 end
