+++ spec/controllers/sponsors_controller_spec.rb
@@ -559,10 +559,10 @@ RSpec.describe SponsorsController, type: :controller do
       let(:petition) { FactoryBot.create(:pending_petition) }
       let(:signature) { FactoryBot.create(:fraudulent_signature, petition: petition, sponsor: true) }
 
-      it "raises an ActiveRecord::RecordNotFound exception" do
+      it "doesn't raise an ActiveRecord::RecordNotFound exception" do
         expect {
           get :verify, id: signature.id, token: signature.perishable_token
-        }.to raise_error(ActiveRecord::RecordNotFound)
+        }.not_to raise_error
       end
     end
 
@@ -570,10 +570,10 @@ RSpec.describe SponsorsController, type: :controller do
       let(:petition) { FactoryBot.create(:pending_petition) }
       let(:signature) { FactoryBot.create(:invalidated_signature, petition: petition, sponsor: true) }
 
-      it "raises an ActiveRecord::RecordNotFound exception" do
+      it "doesn't raise an ActiveRecord::RecordNotFound exception" do
         expect {
           get :verify, id: signature.id, token: signature.perishable_token
-        }.to raise_error(ActiveRecord::RecordNotFound)
+        }.not_to raise_error
       end
     end
 
@@ -932,10 +932,10 @@ RSpec.describe SponsorsController, type: :controller do
       let(:petition) { FactoryBot.create(:pending_petition) }
       let(:signature) { FactoryBot.create(:fraudulent_signature, petition: petition, sponsor: true) }
 
-      it "raises an ActiveRecord::RecordNotFound exception" do
+      it "doesn't raise an ActiveRecord::RecordNotFound exception" do
         expect {
           get :signed, id: signature.id
-        }.to raise_error(ActiveRecord::RecordNotFound)
+        }.not_to raise_error
       end
     end
 
@@ -943,10 +943,10 @@ RSpec.describe SponsorsController, type: :controller do
       let(:petition) { FactoryBot.create(:pending_petition) }
       let(:signature) { FactoryBot.create(:invalidated_signature, petition: petition, sponsor: true) }
 
-      it "raises an ActiveRecord::RecordNotFound exception" do
+      it "doesn't raise an ActiveRecord::RecordNotFound exception" do
         expect {
           get :signed, id: signature.id
-        }.to raise_error(ActiveRecord::RecordNotFound)
+        }.not_to raise_error
       end
     end
